profilechange
profilechage

+++ Team5Spring/src/main/java/spring/form/acorn/ChefController.java
@@ -5,6 +5,7 @@ import java.io.File;
 import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,6 +29,8 @@ public class ChefController {
 	
 	@Autowired
 	private ChefDaoInter dao;
+	@Autowired
+	JavaMailSender mailSender;
 	
 	//아이디 유효성 검사
 	@PostMapping("/chef/checkid")
@@ -102,8 +105,8 @@ public class ChefController {
 		if(pass!=null) {
 			String subject = "슬기로운 냉장고 비밀번호 변경안내";
 			String newpass = dao.RandomPass();
-			String content = "비밀번호가 변경되었습니다. 새로운 비밀번호: "+newpass;
-			MailSend mail = new MailSend();
+			String content = "비밀번호가 변경되었습니다.\r새로운 비밀번호: "+newpass;
+			MailSend mail = new MailSend(mailSender);
 			int checkemail= mail.MailGo(email, subject, content);
 			dao.updatePass(pass, newpass, email);
 			if(checkemail==1)
@@ -121,21 +124,30 @@ public class ChefController {
 	
 	//일반정보수정
 	@RequestMapping(value="/chef/mod", consumes = {"multipart/form-data"}, method = RequestMethod.POST)
-	public int mod(MultipartHttpServletRequest request, @ModelAttribute("ChefDto") ChefDto dto, BindingResult result) {
+	public int mod(MultipartHttpServletRequest request, @RequestParam int change, @ModelAttribute("ChefDto") ChefDto dto, BindingResult result) {
 		String preprofile=dao.getprofile(dto.getEmail());
-		if(dto.getProfileimage()==null)
+		if(change==1) {
+			if(dto.getProfileimage()==null) {
+				String path=request.getSession().getServletContext().getRealPath("/WEB-INF/image/profile");
+				File file = new File(path+"\\"+preprofile);
+				if(file.exists())
+					file.delete();
+				dto.setProfile("basic_user.png");				
+			}else {
+				String path=request.getSession().getServletContext().getRealPath("/WEB-INF/image/profile");
+				String fileName = new Date().getTime()+"_"+dto.getProfileimage().getOriginalFilename();
+				dto.setProfile(fileName);
+				SpringFileWrite sfw = new SpringFileWrite();
+				sfw.writeFileRename(dto.getProfileimage(), path, fileName);
+				File file = new File(path+"\\"+preprofile);
+				if(file.exists())
+					file.delete();
+			}
+		}else {
 			//사진이 안바꼈을 경우 기존 프로필 사진 저장해두기
 			dto.setProfile(preprofile);
-		else {
-			String path=request.getSession().getServletContext().getRealPath("/WEB-INF/image/profile");
-			String fileName = new Date().getTime()+"_"+dto.getProfileimage().getOriginalFilename();
-			dto.setProfile(fileName);
-			SpringFileWrite sfw = new SpringFileWrite();
-			sfw.writeFileRename(dto.getProfileimage(), path, fileName);
-			File file = new File(path+"\\"+preprofile);
-			if(file.exists())
-				file.delete();
 		}
+		
 		dao.updateInfo(dto);
 		return 1;
 	}
